29/6

+++ assets/js/tin-tuc-hd.js
@@ -1,13 +1,13 @@
 $(document).ready(function() {
     $('#zoom-in').on('click',function(){
-        var currentFontSize = parseFloat($('#text p').css('font-size'));
-        var currentFontLineHeight = parseFloat($('#text p').css('line-height'));
+        var currentFontSize = parseInt($('#text p').css('font-size'));
+        var currentFontLineHeight = parseInt($('#text p').css('line-height'));
         $('#text p').css('font-size', (currentFontSize + 2) + 'px');
         $('#text p').css('line-height', (currentFontLineHeight + 2) + 'px');
     })
     $('#zoom-out').on('click',function(){
-        var currentFontSize = parseFloat($('#text p').css('font-size'));
-        var currentFontLineHeight = parseFloat($('#text p').css('line-height'));
+        var currentFontSize = parseInt($('#text p').css('font-size'));
+        var currentFontLineHeight = parseInt($('#text p').css('line-height'));
         $('#text p').css('font-size', (currentFontSize - 2) + 'px');
         $('#text p').css('line-height', (currentFontLineHeight - 2) + 'px');
     })
